+++ src/InstancedQuads/Shaders/Fragment.metal
